LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-which-value-expected.txt


PASS Use the after-change value of interpolate-size: numeric-only when starting CSS transition from 100px to min-content
PASS Use the non-animation value of interpolate-size: numeric-only when starting CSS animation from 100px to min-content
PASS Use the after-change value of interpolate-size: numeric-only when starting CSS transition from min-content to 100px
PASS Use the non-animation value of interpolate-size: numeric-only when starting CSS animation from min-content to 100px
FAIL Use the after-change value of interpolate-size: allow-keywords when starting CSS transition from 100px to min-content assert_equals: width after change expected "100px" but got "0px"
FAIL Use the non-animation value of interpolate-size: allow-keywords when starting CSS animation from 100px to min-content assert_equals: width with animation expected "90px" but got "100px"
FAIL Use the after-change value of interpolate-size: allow-keywords when starting CSS transition from min-content to 100px assert_equals: width after change expected "0px" but got "100px"
FAIL Use the non-animation value of interpolate-size: allow-keywords when starting CSS animation from min-content to 100px assert_equals: width with animation expected "10px" but got "0px"
